pull all non-command functions from commands into lib files
don't let completed poms get alerted
