+++ src/pombot/lib/check-timer.js
@@ -4,7 +4,17 @@ import getErrorHandler from './get-error-handler';
 import getTimeString from './get-time-string';
 
 export default function checkTimer(onAlertedCallback, onCompletedCallback) {
-  query.updatePomsSetAlerted({})
+
+  query.updatePomsSetCompleted()
+      .then(completedRes => {
+        completedRes.map(pom => {
+          // complete each pom that has just been set as is_completed=true
+          onCompletedCallback(pom.slack_id);
+        });
+        return completedRes;
+      }).catch(getErrorHandler('timer->updatePomsSetCompleted', 'failed to update poms to be completed'));
+
+  query.updatePomsSetAlerted()
     .then(alertedRes => {
       alertedRes.map(pom => {
         // alert each pom that has just been set as is_alerted=true
@@ -12,13 +22,4 @@ export default function checkTimer(onAlertedCallback, onCompletedCallback) {
       });
       return alertedRes;
     }).catch(getErrorHandler('timer->updatePomsSetAlerted', 'failed to update poms to be alerted'));
-
-  query.updatePomsSetCompleted()
-    .then(completedRes => {
-      completedRes.map(pom => {
-        // complete each pom that has just been set as is_completed=true
-        onCompletedCallback(pom.slack_id);
-      });
-      return completedRes;
-    }).catch(getErrorHandler('timer->updatePomsSetCompleted', 'failed to update poms to be completed'));
 }
